+++ macros/dbt_utils/date_trunc.sql
@@ -1,6 +1,0 @@
-{% macro materialize__date_trunc(datepart, date) %}
-    date_trunc(
-        {{datepart}},
-         cast({{date}} as timestamp)
-    )
-{% endmacro %}