+++ macros/dbt_utils/datediff.sql
@@ -1,5 +1,3 @@
 {% macro materialize__datediff(first_date, second_date, datepart) %}
-
-  {{ return(dbt_utils.postgres__datediff(first_date, second_date, datepart)) }}
-
+    {{ return(dbt_utils.postgres__datediff(first_date, second_date, datepart)) }}
 {% endmacro %}