+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # materialize-dbt-utils Changelog
 
+## 0.2.0 - 2021-02-14
+
+* Fix the `last_day` macro.
+
+* Update documentation to reflect newly supported macros when using
+  dbt-materialize v1.0.0.post2+ and Materialize v0.20.0+.
+
 ## 0.1.0 - 2021-01-31
 
 * Initial release.